Add error handling
[step # 5] Innovation summary:
- Implemented error trapping with output redirection to a log file.
- Added tests for the work of exceptions.

+++ page_loader/loading_handler/downloader.py
@@ -1,13 +1,12 @@
 import os
-from typing import Final
+import sys
+import traceback
 
-from page_loader.loading_handler.file_system_guide import get_file_path, get_dir_path  # noqa: E501
+from page_loader.loading_handler.file_system_guide import DEFAULT_DIR, \
+    get_file_path, get_dir_path
 from page_loader.loading_handler.html_parser import parse_page
 from page_loader.logger import logger, \
-    START_DOWNLOAD, END_DOWNLOAD
-
-
-DEFAULT_DIR: Final[str] = os.getcwd()
+    START_DOWNLOAD, FINISH_DOWNLOAD, NOT_FOUND_DESTINATION
 
 
 def download(url: str, destination: str = DEFAULT_DIR) -> str:
@@ -28,16 +27,24 @@ def download(url: str, destination: str = DEFAULT_DIR) -> str:
     ---
         file_path (str): Full path to the downloaded file.
     '''
-    logger.info(START_DOWNLOAD.format(url, destination))
+    try:
+        if not os.path.exists(destination):
+            raise ValueError(NOT_FOUND_DESTINATION.format(destination))
 
-    file_path = get_file_path(url, destination)
-    dir_path = get_dir_path(url, destination)
+        file_path = get_file_path(url, destination)
+        dir_path = get_dir_path(url, destination)
 
-    html = parse_page(url, dir_path)
+        logger.info(START_DOWNLOAD.format(url, destination))
 
-    with open(file_path, 'w') as file:
-        file.write(html)
+        html = parse_page(url, dir_path)
 
-    logger.info(END_DOWNLOAD.format(file_path, dir_path))
+        with open(file_path, 'w') as file:
+            file.write(html)
+
+    except Exception:
+        logger.error(traceback.format_exc())
+        sys.exit(1)
+
+    logger.info(FINISH_DOWNLOAD.format(file_path, dir_path))
 
     return file_path
